WebGPURenderer: Introduce ShaderLib.
